Load CDK: Remove dead streaming upload code
