+++ airbyte-cdk/bulk/toolkits/load-object-storage/src/main/kotlin/io/airbyte/cdk/load/file/object_storage/ObjectStorageClient.kt
@@ -4,9 +4,7 @@
 
 package io.airbyte.cdk.load.file.object_storage
 
-import io.airbyte.cdk.load.file.StreamProcessor
 import java.io.InputStream
-import java.io.OutputStream
 import kotlinx.coroutines.flow.Flow
 
 interface ObjectStorageClient<T : RemoteObject<*>> {
@@ -18,20 +16,6 @@ interface ObjectStorageClient<T : RemoteObject<*>> {
     suspend fun put(key: String, bytes: ByteArray): T
     suspend fun delete(remoteObject: T)
     suspend fun delete(key: String)
-
-    /**
-     * Streaming upload should provide an [OutputStream] managed within the lifecycle of [block].
-     * The stream should be closed after the block completes, however it should be safe for users of
-     * the stream to close early (some writers do this by default, especially those that write whole
-     * files). Specifically, the method should guarantee that no operations will be performed on the
-     * stream after [block] completes.
-     */
-    suspend fun <V : OutputStream> streamingUpload(
-        key: String,
-        metadata: Map<String, String> = emptyMap(),
-        streamProcessor: StreamProcessor<V>? = null,
-        block: suspend (OutputStream) -> Unit
-    ): T
 
     /** Experimental sane replacement interface */
     suspend fun startStreamingUpload(
